+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change the type of `Extensions` from `Mapping[Str, Any]` to `MutableMapping[Str, Any]`. (#762)
 - Handle HTTP/1.1 half-closed connections gracefully. (#641)
 - Drop Python 3.7 support. (#727)
-- Added `httpcore.ProxyMode` to make HTTP proxy mode configurable. You can now force proxy connection to be TUNNEL or GATEWAY. 
+- Added `httpcore.ProxyMode` to make HTTP proxy mode configurable. You can now force proxy connection to be TUNNEL or FORWARD. 
 
 ## 0.17.3 (July 5th, 2023)