+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 0.18.0
+## 0.18.0 (September 8th, 2023)
 
 - Add support for HTTPS proxies. (#745, # 786)
 - Handle `sni_hostname` extension with SOCKS proxy. (#774)